refactor(#114): useRequireAuth 관련 리팩토링
