--- src/hooks/useAuth.tsx
@@ -1,17 +1,0 @@
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-
-export const useRequireAuth = () => {
-  const router = useRouter();
-
-  useEffect(() => {
-    // const token = getToken();
-    // const isFirstLogin = localStorage.getItem("isFirstLogin");
-    // 토큰이 없으면 로그인 페이지로 리다이렉트
-    // if (!token) {
-    //   router.push("/");
-    // } else if (isFirstLogin) {
-    //   router.push("/first/step1");
-    // }
-  }, []);
-};
